refactor folder structure and add readme
