+++ README.md
@@ -1,1 +1,15 @@
-# Virtual_Dragon_Boat_Coach
+# Virtual Dragon Boat Coach
+
+A smart coach application that analyzes dragon boat paddling technique using pose estimation and provide feedback for improving paddling form to prevent injuries.
+
+### Required Credentials
+The `capstone_sa.json` credential file is required for Google Drive functionality. This file contains sensitive API credentials and is not included in the repository. Please contact one of the following authors to obtain access:
+- James Park (parkg10@mcmaster.ca)
+- Iain Slim (slimi@mcmaster.ca)
+- Logan Yuan ()yuanq10@mcmaster.ca
+- Richard Park (parks110@mcmaster.ca)
+
+
+### Dependencies
+```bash
+pip install -r requirements.txt
