+++ requirements.txt
@@ -1,0 +1,6 @@
+ultralytics
+opencv-python
+numpy
+matplotlib
+scipy
+google-api-python-client
